move mark evaluator into mark package
